+++ constants/constants.go
@@ -2,10 +2,15 @@ package constants
 
 import (
 	"fmt"
+	"io/ioutil"
+	"path/filepath"
 
+	"github.com/abdfnx/tran/dfs"
 	"github.com/charmbracelet/lipgloss"
 	"github.com/evertras/bubble-table/table"
 )
+
+const SMUI_PORT = "3750"
 
 const PRIMARY_COLOR = "#1163E6"
 const SECONDARY_COLOR = "#B4B4B4"
@@ -81,3 +86,33 @@ $ secman delete --[PASSWORD_TYPE] <PASSWORD_NAME>
 
 $ secman generate
 `, Logo("Initialize ~/.secman"), Logo("Authorize With Secman"), Logo("Insert a New Password"), Logo("List All Passwords"), Logo("Read a Password"), Logo("Edit a Password"), Logo("Delete"), Logo("Generate"))
+
+var (
+	homeDir, homeErr = dfs.GetHomeDirectory()
+	DotSecmanPath = filepath.Join(homeDir, ".secman")
+	SecmanConfigPath = filepath.Join(DotSecmanPath, "secman.json")
+	SMUIPath = filepath.Join(DotSecmanPath, "ui")
+	SMUIIndexPath = filepath.Join(SMUIPath, "index.html")
+	secmanConfig, SmErr = ioutil.ReadFile(SecmanConfigPath)
+)
+
+func SecmanConfig() []byte {
+	if SmErr != nil {
+		return []byte("")
+	}
+
+	return secmanConfig
+}
+
+var (
+	checkmarkChar = lipgloss.NewStyle().
+		SetString("✔").
+		Foreground(lipgloss.Color(GREEN_COLOR))
+
+	xChar = lipgloss.NewStyle().
+		SetString("✘").
+		Foreground(lipgloss.Color(RED_COLOR))
+
+	Checkmark = "[" + checkmarkChar.String() + "] "
+	X = "[" + xChar.String() + "] "
+)